exam/consumer/requirements.txt
numpy == 1.23.4;
scikit-learn==1.5.1;
kafka-python
